Save output log
