+++ tpcbench-run/src/main/scala/sc/RunResult.scala
@@ -1,0 +1,15 @@
+package sc
+
+final case class RunResult(
+  succeeded: Boolean,
+  wallTime: Int,
+  queryTime: Int,
+  appUrl: String,
+  traceResults: String,
+  logOutput: String
+) {}
+object RunResult {
+  val fieldNames: List[String] = {
+    classOf[RunResult].getDeclaredFields.map(_.getName).toList
+  }
+}
