+++ tpcbench-run/src/test/scala/sc/TpcBenchSpec.scala
@@ -2,7 +2,7 @@ package sc
 
 import cats.effect.unsafe.implicits.global
 import org.scalatest.freespec.AnyFreeSpec
-import sc.RunOptions.{Log4jFile, RunResult}
+import sc.RunOptions.Log4jFile
 
 import java.nio.file.Files
 
@@ -14,12 +14,8 @@ final class TpcBenchSpec extends AnyFreeSpec {
     assert(RunOptions.fieldNames.contains("name"))
   }
 
-  import doobie._
-  import doobie.implicits._
-  import cats._
-  import cats.data._
   import cats.effect.IO
-  import cats.implicits._
+  import doobie._
 
   val xa = Transactor
     .fromDriverManager[IO]("org.sqlite.JDBC", "jdbc:sqlite:test.db", "", "")
@@ -35,7 +31,8 @@ final class TpcBenchSpec extends AnyFreeSpec {
         wallTime = 1,
         queryTime = 123,
         traceResults = "KK",
-        appUrl = "abc"
+        appUrl = "abc",
+        logOutput = (0 to 40).map(l => s"s$l").mkString("\n")
       )
     )).unsafeRunSync()
   }
